requirements.txt
astroid==2.5.6
certifi==2021.5.30
chardet==4.0.0
discord-webhook==0.14.0
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
pylint==2.8.3
PySocks==1.7.1
python-dotenv==0.18.0
requests==2.25.1
requests-oauthlib==1.3.0
Rx==3.2.0
six==1.16.0
toml==0.10.2
tweepy==3.10.0
twitch-python==0.0.19
urllib3==1.26.6
wrapt==1.12.1
instabot==0.117.0
